[Main] Support object reuse (#2975)
* support object reuse

* fix formatting

---------
